reorganized directories and solved "key" error on console
