+++ client/src/App.jsx
@@ -3,10 +3,8 @@ import React from 'react';
 import Overview from './product-details/Overview';
 import RatingsReviews from './RatingsReviews/RatingsReviews';
 import QuestionAnswer from './questionsAnswers/QuestionAnswers';
-import RelatedItems from './related-items/related-items.jsx';
-import Overview from './product-details/Overview.jsx';
-import { CurrentProductProvider } from './context.jsx';
-
+import RelatedItems from './related-items/related-items';
+import { CurrentProductProvider } from './context';
 
 function App() {
   return (
