+++ client/src/index.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App.jsx';
-
-
+import App from './App';
 
 const root = createRoot(document.getElementById('root'));
