+++ client/src/questionsAnswers/buttons/AddQuestionForm.jsx
@@ -1,18 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
-import FormInput from './FormInput';
+import FormInput from '../lib/FormInput';
 
-function AddAnswerForm({ show, setShowModal, questionId, questionBody, submitHandler, productName }) {
+function AddQuestionForm({ show, setShowModal, questionId, submitHandler, productName }) {
   if (!show) return null;
-  const reader = new FileReader();
-
   const [isFormValid, setIsFormValid] = useState(true);
-  const [imageInvalid, setImageInvalid] = useState(false);
   const [emptyFields, setEmptyFields] = useState();
   const [invalidMessage, setInvalidMessage] = useState([]);
   const [formValue, setFormValue] = useState({});
 
-  const inputChangeHandler = (e) => {
+  const onChange = (e) => {
     const { name, value } = e.target;
     setFormValue({
       ...formValue,
@@ -20,68 +17,47 @@ function AddAnswerForm({ show, setShowModal, questionId, questionBody, submitHan
     });
   };
 
-  const fileChangeHandler = (e) => {
-    const uploadedFile = e.target.files[0];
-
-    reader.onload = (event) => {
-      const img = new Image();
-      img.onload = () => {
-        const newImages = formValue.photos ? [...formValue.photos] : [];
-        newImages.push({
-          name: uploadedFile[0].name,
-          url: URL.createObjectURL(uploadedFile[0]),
-        });
-        setFormValue({
-          ...formValue,
-          photos: newImages,
-        });
-      };
-      img.onerror = () => {
-        setImageInvalid(true);
-      };
-      img.src = event.target.result;
-    };
-    reader.readAsDataURL(uploadedFile);
-  };
-
   const inputs = [
     {
       config: {
-        label: 'Your Answer',
+        label: 'Your Question',
         type: 'textarea',
         name: 'body',
-        value: formValue.body,
+        value: formValue.body || '',
         placeholder: '',
+        mandatory: 'true',
         maxLength: 1000,
-        mandatory: true,
+        onChange,
       },
-      changeHandler: inputChangeHandler,
+      comment: 'For privacy reasons, do not use your full name or email address',
+      changeHandler: onChange,
     },
     {
       config: {
         label: 'What is your nickname',
-        type: 'text',
+        type: 'username',
         name: 'name',
-        value: formValue.name,
-        placeholder: 'Example: jack543!',
+        value: formValue.name || '',
+        placeholder: 'Example: jackson11!',
+        mandatory: 'true',
         maxLength: 60,
-        mandatory: true,
+        onChange,
       },
-      comment: 'For privacy reasons, do not use your full name or email address',
-      changeHandler: inputChangeHandler,
+      comment: 'For authentication reasons, you will not be emailed',
+      changeHandler: onChange
     },
     {
       config: {
         label: 'Your email',
         type: 'email',
         name: 'email',
-        value: formValue.email,
+        value: formValue.email || '',
         placeholder: 'Example: jack@email.com',
+        mandatory: 'true',
         maxLength: 60,
-        mandatory: true,
+        onChange,
       },
-      comment: 'For authentication reasons, you will not be emailed',
-      changeHandler: inputChangeHandler,
+      changeHandler: onChange,
     },
   ];
 
@@ -93,7 +69,6 @@ function AddAnswerForm({ show, setShowModal, questionId, questionBody, submitHan
   const validateForm = () => {
     let result = true;
     const invalid = [];
-    const newInvalidMessage = [];
     inputs.forEach(({ config }) => {
       const target = config.name;
       if (Boolean(config.mandatory) && !formValue[target]) {
@@ -109,11 +84,12 @@ function AddAnswerForm({ show, setShowModal, questionId, questionBody, submitHan
         result = false;
       }
       if (config.type === 'email' && !validateEmail(formValue[config.name])) {
-        newInvalidMessage.push('Email is invalid.');
+        setInvalidMessage([
+          'Email is invalid.',
+        ]);
         result = false;
       }
     });
-    setInvalidMessage(newInvalidMessage);
     setEmptyFields(invalid);
     setIsFormValid(result);
     return result;
@@ -140,14 +116,13 @@ function AddAnswerForm({ show, setShowModal, questionId, questionBody, submitHan
     <Modal>
       <PopupForm id={`${questionId}-popup`}>
         <Header>
-          <Title>Submit your Answer</Title>
-          <Subtitle>{`${productName}:`}</Subtitle>
-          <Subtitle>{questionBody}</Subtitle>
+          <Title>Ask Your Question</Title>
+          <Subtitle>{'About the ' + productName}</Subtitle>
         </Header>
         {!isFormValid
         && (
         <Invalid>
-          {emptyFields.length > 0
+          {Boolean(emptyFields.length)
           && (
             <li>
               You must enter the following:
@@ -160,45 +135,16 @@ function AddAnswerForm({ show, setShowModal, questionId, questionBody, submitHan
         </Invalid>
         )}
 
-        {inputs.map(({ config, comment, changeHandler }) => (
+        {inputs.map(({config, comment, changeHandler}) => (
           <div>
             <FormInput
-              key={config.label}
+              key={config.name}
               attribute={config}
               changeHandler={changeHandler}
             />
             {comment && `*${comment}`}
           </div>
         ))}
-        <div>
-          <FormInput
-            attribute={{
-              label: 'Upload your photos (choose up to 5 photos)',
-              type: 'file',
-              name: 'photos',
-              placeholder: '',
-            }}
-            changeHandler={fileChangeHandler}
-          />
-          <ThumbnailContainer>
-            {imageInvalid && <div style={{ color: 'red' }}>Invalid image content.</div> }
-            {!formValue.photos && <p>No files selected.</p>}
-            {formValue.photos && formValue.photos.map((photo) => (
-              <FlexRowDiv>
-                <div style={{ width: '50%' }}>
-                  <Thumbnail
-                    className="obj"
-                    key={photo.url}
-                    src={photo.url}
-                  />
-                </div>
-                <div style={{ width: '50%' }}>
-                  {photo.name}
-                </div>
-              </FlexRowDiv>
-            ))}
-          </ThumbnailContainer>
-        </div>
         <DivButton className="form-buttons">
           <ButtonStyled type="submit" onClick={handleSubmit} data-testid="form-button-test">
             Submit
@@ -212,7 +158,7 @@ function AddAnswerForm({ show, setShowModal, questionId, questionBody, submitHan
   );
 }
 
-export default AddAnswerForm;
+export default AddQuestionForm;
 
 const Modal = styled.div`
   position: fixed;
@@ -261,25 +207,9 @@ const Subtitle = styled.div`
   padding: 2px 0;
 `;
 
+
 const Invalid = styled.ul`
   color: red;
-`;
-
-const ThumbnailContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: 10px 0;
-`;
-
-const Thumbnail = styled.img`
-  max-width:100px;
-  max-height: 100px;
-  padding-right:10px;
-`;
-
-const FlexRowDiv = styled.div`
-  display: flex;
-  flex-direction: row;
 `;
 
 const PaddedDiv = styled.div`
